ml/requirements.txt
yt-dlp
pydub
openai-whisper
torch
numpy
tqdm
more-itertools
transformers
scikit-learn
bertopic
sentence-transformers
umap-learn
hdbscan
nltk
pandas